Add files via upload
J. Peeples, D. Suen, A. Zare, and J. Keller, "Possibilistic fuzzy local information C-means with automated feature selection for seafloor segmentation" in Proc. SPIE 10628. Detection and Sensing of Mines, Explosive Objects, and Obscured Targets XXIII,2018.
Abstract: The Possibilistic Fuzzy Local Information C-Means (PFLICM) method is presented as a technique to segment side-look synthetic aperture sonar (SAS) imagery into distinct regions of the sea-floor. In this work, we investigate and present the results of an automated feature selection approach for SAS image segmentation. The chosen features and resulting segmentation from the image will be assessed based on a select quantitative clustering validity criterion and the subset of the features that reach a desired threshold will be used for the segmentation process.